OpenCV.Calib3d

Documentation

data FundamentalMatMethod

Constructors

FM_7Point

FM_8Point

FM_Ransac !(Maybe Double) !(Maybe Double)

FM_Lmeds !(Maybe Double)

data WhichImage

Constructors

Image1

Image2

findFundamentalMat

Arguments

:: IsPoint2 point2 CDouble

=> Vector (point2 CDouble)
Points from the first image.

-> Vector (point2 CDouble)
Points from the second image.

-> FundamentalMatMethod

-> CvExcept (Maybe (Mat (S '[D, S 3]) (S 1) (S Double), Mat (S '[D, D]) (S 1) (S Word8)))

Calculates a fundamental matrix from the corresponding points in two images

The minimum number of points required depends on the FundamentalMatMethod.

FM_7Point: N == 7

FM_8Point: N >= 8

FM_Ransac: N >= 15

FM_Lmeds: N >= 8

With 7 points the FM_7Point method is used, despite the given method.

With more than 7 points the FM_7Point method will be replaced by the FM_8Point method.

Between 7 and 15 points the FM_Ransac method will be replaced by the FM_Lmeds method.

With the FM_7Point method and with 7 points the result can contain up to 3 matrices, resulting in either 3, 6 or 9 rows. This is why the number of resulting rows in tagged as Dynamic. For all other methods the result always contains 3 rows.

OpenCV Sphinx doc

computeCorrespondEpilines

Arguments

:: IsPoint2 point2 CDouble

=> Vector (point2 CDouble)
Points.

-> WhichImage
Image which contains the points.

-> Mat (ShapeT '[3, 3]) (S 1) (S Double)
Fundamental matrix.

-> CvExcept (Mat (S '[D, S 1]) (S 3) (S Double))

For points in an image of a stereo pair, computes the corresponding epilines in the other image

OpenCV Sphinx doc