OpenCV.Core.Types.Matx

Documentation

data Matx depth m n

type family MatxDepth (a :: *) :: *

Calculate the depth for types which can be converted to or from Matx

type family MatxDimM (a :: *) :: Nat

Calculate the first dimension (called m) for types which can be converted to or from Matx

type family MatxDimN (a :: *) :: Nat

Calculate the second dimension (called n) for types which can be converted to or from Matx

class ToMatx a where

Minimal complete definition

toMatx

Methods

toMatx :: a -> Matx (MatxDepth a) (MatxDimM a) (MatxDimN a)

class FromMatx a where

Minimal complete definition

fromMatx

Methods

fromMatx :: Matx (MatxDepth a) (MatxDimM a) (MatxDimN a) -> a

data Vec depth dim

type family VecDepth (a :: *) :: *

Calculate the depth for types which can be converted to or from Vec

type family VecDim (a :: *) :: Nat

Calculate the dimension for types which can be converted to or from Vec

class ToVec a where

Minimal complete definition

toVec

Methods

toVec :: a -> Vec (VecDepth a) (VecDim a)

class FromVec a where

Minimal complete definition

fromVec

Methods

fromVec :: Vec (VecDepth a) (VecDim a) -> a

type Vec2i = Vec Int32 2

type Vec2f = Vec CFloat 2

type Vec2d = Vec CDouble 2

type Vec3i = Vec Int32 3

type Vec3f = Vec CFloat 3

type Vec3d = Vec CDouble 3

type Vec4i = Vec Int32 4

type Vec4f = Vec CFloat 4

type Vec4d = Vec CDouble 4

newVec2i :: V2 Int32 -> IO Vec2i

newVec2f :: V2 CFloat -> IO Vec2f

newVec2d :: V2 CDouble -> IO Vec2d

newVec3i :: V3 Int32 -> IO Vec3i

newVec3f :: V3 CFloat -> IO Vec3f

newVec3d :: V3 CDouble -> IO Vec3d

newVec4i :: V4 Int32 -> IO Vec4i

newVec4f :: V4 CFloat -> IO Vec4f

newVec4d :: V4 CDouble -> IO Vec4d