OpenCV.ImgProc.StructuralAnalysis

Documentation

contourArea

Arguments

:: Convert point2f Point2f

=> Vector point2f
Input vector of 2D points (contour vertices).

-> Bool
Oriented area flag. If it is true, the function returns a signed area value, depending on the contour orientation (clockwise or counter-clockwise). Using this feature you can determine orientation of a contour by taking the sign of an area.

-> CvExcept Double

Calculates a contour area.

The function computes a contour area. Similarly to moments, the area is computed using the Green formula. Thus, the returned area and the number of non-zero pixels, if you draw the contour using drawContours or fillPoly, can be different. Also, the function will most certainly give a wrong results for contours with self-intersections.

OpenCV Sphinx doc

pointPolygonTest

Arguments

:: (Convert contourPoint2f Point2f, Convert testPoint2f Point2f)

=> Vector contourPoint2f
Contour.

-> testPoint2f
Point tested against the contour.

-> Bool
If true, the function estimates the signed distance from the point to the nearest contour edge. Otherwise, the function only checks if the point is inside a contour or not.

-> CvExcept Double

Performs a point-in-contour test.

The function determines whether the point is inside a contour, outside, or lies on an edge (or coincides with a vertex). It returns positive (inside), negative (outside), or zero (on an edge) value, correspondingly. When measureDist=false , the return value is +1, -1, and 0, respectively. Otherwise, the return value is a signed distance between the point and the nearest contour edge.

OpenCV Sphinx doc