OpenCV.Internal.Core.Types.Mat

Matrix

newtype Mat shape channels depth

Constructors

Mat

Fields

unMat :: ForeignPtr (C (Mat shape channels depth))

typeCheckMat

Arguments

:: (ToShapeDS (Proxy shape), ToChannelsDS (Proxy channels), ToDepthDS (Proxy depth))

=> Mat shape channels depth
The matrix to be checked.

-> [CoerceMatError]
Error messages.

Tests whether a Mat is deserving of its type level attributes

Checks if the properties encoded in the type of a Mat correspond to the value level representation. For each property that does not hold this function will produce an error message. If everything checks out it will produce an empty list.

The following properties are checked:

Dimensionality

Size of each dimension

Number of channels

Depth (data type of elements)

If a property is explicitly encoded as statically unknown (Dynamic) it will not be checked.

relaxMat

Arguments

:: (MayRelax shapeIn shapeOut, MayRelax channelsIn channelsOut, MayRelax depthIn depthOut)

=> Mat shapeIn channelsIn depthIn
Original Mat.

-> Mat shapeOut channelsOut depthOut
Mat with relaxed constraints.

Relaxes the type level constraints

Only identical or looser constraints are allowed. For tighter constraints use coerceMat.

This allows you to 'forget' type level guarantees for zero cost. Similar to unsafeCoerceMat, but totally safe.

Identical

a to b with a ~ b

Looser

('S a) to 'D or ('S a) to ('S b) with MayRelax a b

Tighter

'D to ('S a)

coerceMat

Arguments

:: (ToShapeDS (Proxy shapeOut), ToChannelsDS (Proxy channelsOut), ToDepthDS (Proxy depthOut))

=> Mat shapeIn channelsIn depthIn

-> CvExcept (Mat shapeOut channelsOut depthOut)

unsafeCoerceMat :: Mat shapeIn channelsIn depthIn -> Mat shapeOut channelsOut depthOut

keepMatAliveDuring :: Mat shape channels depth -> IO a -> IO a

Similar to withPtr in that it keeps the ForeignPtr alive during the execution of the given action but it doesn't extract the Ptr from the ForeignPtr.

newEmptyMat :: IO (Mat (S '[]) (S 1) (S Word8))

newMat

Arguments

:: (ToShape shape, ToChannels channels, ToDepth depth, ToScalar scalar)

=> shape

-> channels

-> depth

-> scalar

-> CvExceptT IO (Mat (ShapeT shape) (ChannelsT channels) (DepthT depth))

withMatData

Arguments

:: Mat shape channels depth

-> ([CSize] -> Ptr Word8 -> IO a)

-> IO a

matElemAddress :: Ptr Word8 -> [Int] -> [Int] -> Ptr a

mkMat

Arguments

:: (ToShape shape, ToChannels channels, ToDepth depth, ToScalar scalar)

=> shape

-> channels

-> depth

-> scalar

-> CvExcept (Mat (ShapeT shape) (ChannelsT channels) (DepthT depth))

cloneMat :: Mat shape channels depth -> Mat shape channels depth

Mutable matrix

typeCheckMatM

Arguments

:: (ToShapeDS (Proxy shape), ToChannelsDS (Proxy channels), ToDepthDS (Proxy depth))

=> Mut (Mat shape channels depth) s
The matrix to be checked.

-> [CoerceMatError]
Error messages.

relaxMatM

Arguments

:: (MayRelax shapeIn shapeOut, MayRelax channelsIn channelsOut, MayRelax depthIn depthOut)

=> Mut (Mat shapeIn channelsIn depthIn) s
Original Mat.

-> Mut (Mat shapeOut channelsOut depthOut) s
Mat with relaxed constraints.

coerceMatM

Arguments

:: (ToShapeDS (Proxy shapeOut), ToChannelsDS (Proxy channelsOut), ToDepthDS (Proxy depthOut))

=> Mut (Mat shapeIn channelsIn depthIn) s

-> CvExcept (Mut (Mat shapeOut channelsOut depthOut) s)

unsafeCoerceMatM :: Mut (Mat shapeIn channelsIn depthIn) s -> Mut (Mat shapeOut channelsOut depthOut) s

mkMatM

Arguments

:: (PrimMonad m, ToShape shape, ToChannels channels, ToDepth depth, ToScalar scalar)

=> shape

-> channels

-> depth

-> scalar

-> CvExceptT m (Mut (Mat (ShapeT shape) (ChannelsT channels) (DepthT depth)) (PrimState m))

createMat

Arguments

:: (forall s. CvExceptT (ST s) (Mut (Mat shape channels depth) s))

-> CvExcept (Mat shape channels depth)

withMatM

Arguments

:: (ToShape shape, ToChannels channels, ToDepth depth, ToScalar scalar)

=> shape

-> channels

-> depth

-> scalar

-> (forall s. Mut (Mat (ShapeT shape) (ChannelsT channels) (DepthT depth)) (PrimState (ST s)) -> CvExceptT (ST s) ())

-> CvExcept (Mat (ShapeT shape) (ChannelsT channels) (DepthT depth))

cloneMatM :: PrimMonad m => Mat shape channels depth -> m (Mat shape channels depth)

Meta information

data MatInfo

Constructors

MatInfo

Fields

miShape :: ![Int32]

miDepth :: !Depth

miChannels :: !Int32

matInfo :: Mat shape channels depth -> MatInfo

data Depth

Constructors

Depth_8U

Depth_8S

Depth_16U

Depth_16S

Depth_32S

Depth_32F

Depth_64F

Depth_USRTYPE1

marshalDepth :: Depth -> Int32

unmarshalDepth :: Int32 -> Depth

marshalFlags

Arguments

:: Depth

-> Int32
Number of channels

-> Int32

unmarshalFlags :: Int32 -> (Depth, Int32)

type family ShapeT (a :: ka) :: DS [DS Nat] where ...

Equations

ShapeT [Int32] = D

ShapeT (Vector Int32) = D

ShapeT (x ::: xs) = S (DSNats (x ::: xs))

ShapeT (xs :: [Nat]) = S (DSNats xs)

ShapeT (Proxy a) = ShapeT a

type ChannelsT a = DSNat a

type family DepthT a :: DS * where ...

Equations

DepthT Depth = D

DepthT (proxy d) = S d

class ToShape a where

Minimal complete definition

toShape

Methods

toShape :: a -> Vector Int32

class ToShapeDS a where

Minimal complete definition

toShapeDS

Methods

toShapeDS :: a -> DS [DS Int32]

type ToChannels a = ToInt32 a

toChannels :: ToInt32 a => a -> Int32

type ToChannelsDS a = ToNatDS a

toChannelsDS :: ToChannelsDS a => a -> DS Int32

class ToDepth a where

Minimal complete definition

toDepth

Methods

toDepth :: a -> Depth

class ToDepthDS a where

Minimal complete definition

toDepthDS

Methods

toDepthDS :: a -> DS Depth

Orphan instances

PlacementNew C'Mat

Methods

placementNew :: Ptr C'Mat -> Ptr C'Mat -> IO ()

placementDelete :: Ptr C'Mat -> IO ()