OpenCV.Video.MotionAnalysis

BackgroundSubtractor

class BackgroundSubtractor a m where

Methods

bgSubApply

Arguments

:: a

-> Double
The value between 0 and 1 that indicates how fast the background model is learnt. Negative parameter value makes the algorithm to use some automatically chosen learning rate. 0 means that the background model is not updated at all, 1 means that the background model is completely reinitialized from the last frame.

-> Mat (S `[h, w]`) channels depth
Next video frame.

-> m (Mat (S `[h, w]`) (S 1) (S Word8))
The output foreground mask as an 8-bit binary image.

getBackgroundImage

Arguments

:: a

-> m (Mat (S `[h, w]`) channels depth)
The output background image.

Background subtractors

data BackgroundSubtractorMOG2 s

data BackgroundSubtractorKNN s

newBackgroundSubtractorKNN

Arguments

:: PrimMonad m

=> Maybe Int32
Length of the history.

-> Maybe Double
Threshold on the squared distance between the pixel and the sample to decide whether a pixel is close to that sample. This parameter does not affect the background update.

-> Maybe Bool
If True, the algorithm will detect shadows and mark them. It decreases the speed a bit, so if you do not need this feature, set the parameter to False.

-> m (BackgroundSubtractorKNN (PrimState m))

newBackgroundSubtractorMOG2

Arguments

:: PrimMonad m

=> Maybe Int32
Length of the history.

-> Maybe Double
Threshold on the squared Mahalanobis distance between the pixel and the model to decide whether a pixel is well described by the background model. This parameter does not affect the background update.

-> Maybe Bool
If True, the algorithm will detect shadows and mark them. It decreases the speed a bit, so if you do not need this feature, set the parameter to False.

-> m (BackgroundSubtractorMOG2 (PrimState m))