OpenCV.VideoIO.Types

VideoCodecs

newtype FourCC

Constructors

FourCC

Fields

unFourCC :: Int32

data VideoCaptureProperties

Constructors

VideoCapPropPosMsec
Current position of the video file in milliseconds.

VideoCapPropPosFrames
0-based index of the frame to be decoded/captured next.

VideoCapPropPosAviRatio
Relative position of the video file: 0=start of the film, 1=end of the film.

VideoCapPropFrameWidth
Width of the frames in the video stream.

VideoCapPropFrameHeight
Height of the frames in the video stream.

VideoCapPropFps
Frame rate.

VideoCapPropFourCc
4-character code of codec.

VideoCapPropFrameCount
Number of frames in the video file.

VideoCapPropFormat
Format of the Mat objects returned by VideoCapture::retrieve().

VideoCapPropMode
Backend-specific value indicating the current capture mode.

VideoCapPropBrightness
Brightness of the image (only for cameras).

VideoCapPropContrast
Contrast of the image (only for cameras).

VideoCapPropSaturation
Saturation of the image (only for cameras).

VideoCapPropHue
Hue of the image (only for cameras).

VideoCapPropGain
Gain of the image (only for cameras).

VideoCapPropExposure
Exposure (only for cameras).

VideoCapPropConvertRgb
Boolean flags indicating whether images should be converted to RGB.

VideoCapPropWhiteBalanceBlueU
Currently unsupported.

VideoCapPropRectification
Rectification flag for stereo cameras (note: only supported by DC1394 v 2.x backend currently).

VideoCapPropMonochrome

VideoCapPropSharpness

VideoCapPropAutoExposure
DC1394: exposure control done by camera, user can adjust reference level using this feature.

VideoCapPropGamma

VideoCapPropTemperature

VideoCapPropTrigger

VideoCapPropTriggerDelay

VideoCapPropWhiteBalanceRedV

VideoCapPropZoom

VideoCapPropFocus

VideoCapPropGuid

VideoCapPropIsoSpeed

VideoCapPropBacklight

VideoCapPropPan

VideoCapPropTilt

VideoCapPropRoll

VideoCapPropIris

VideoCapPropSettings
Pop up video/camera filter dialog (note: only supported by DSHOW backend currently. Property value is ignored)

VideoCapPropBuffersize

VideoCapPropAutofocus

VideoCapPropInt !Int32
Any property we need. Meaning of this property depends on the backend.

data VideoCaptureAPI

Constructors

VideoCapAny
Auto detect == 0

VideoCapVfw
Video For Windows (platform native)

VideoCapV4l
V4L/V4L2 capturing support via libv4l

VideoCapV4l2
Same as CAP_V4L

VideoCapFirewire
IEEE 1394 drivers

VideoCapFireware
Same as CAP_FIREWIRE

VideoCapIeee1394
Same as CAP_FIREWIRE

VideoCapDc1394
Same as CAP_FIREWIRE

VideoCapCmu1394
Same as CAP_FIREWIRE

VideoCapQt
QuickTime

VideoCapUnicap
Unicap drivers

VideoCapDshow
DirectShow (via videoInput)

VideoCapPvapi
PvAPI, Prosilica GigE SDK

VideoCapOpenni
OpenNI (for Kinect)

VideoCapOpenniAsus
OpenNI (for Asus Xtion)

VideoCapAndroid
Android - not used

VideoCapXiapi
XIMEA Camera API

VideoCapAvfoundation
AVFoundation framework for iOS (OS X Lion will have the same API)

VideoCapGiganetix
Smartek Giganetix GigEVisionSDK

VideoCapMsmf
Microsoft Media Foundation (via videoInput)

VideoCapWinrt
Microsoft Windows Runtime using Media Foundation

VideoCapIntelperc
Intel Perceptual Computing SDK

VideoCapOpenni2
OpenNI2 (for Kinect)

VideoCapOpenni2Asus
OpenNI2 (for Asus Xtion and Occipital Structure sensors)

VideoCapGphoto2
gPhoto2 connection

VideoCapGstreamer
GStreamer

VideoCapFfmpeg
Open and record video file or stream using the FFMPEG library

VideoCapImages
Image Sequence (e.g. img_%02d.jpg)