OpenCV.VideoIO.VideoWriter

Documentation

data VideoWriter

data VideoWriterSink

Constructors

VideoFileSink' !VideoFileSink

data VideoFileSink

Constructors

VideoFileSink

Fields

vfsFilePath :: !FilePath

vfsFourCC :: !FourCC

vfsFps :: !Double

vfsFrameDims :: !(Int32, Int32)

videoWriterOpen :: VideoWriterSink -> IO VideoWriter

The API might change in the future, but currently we can:

Open/create a new file:

  wr <- videoWriterOpen $ VideoFileSink'
     (VideoFileSink
        "tst.MOV"
        "avc1"
        30
        (3840, 2160)
     )

Now, we can write some frames, but they need to have exactly the same size as the one we have opened with:

  exceptErrorIO $ videoWriterWrite wr img

We need to close at the end or it will not finalize the file:

  exceptErrorIO $ videoWriterRelease wr

videoWriterRelease :: VideoWriter -> CvExceptT IO ()

videoWriterIsOpened :: VideoWriter -> IO Bool

videoWriterWrite :: VideoWriter -> Mat (S '[D, D]) D D -> CvExceptT IO ()