OpenCV.Core.Types

Mutable values

data Mut a s

Wrapper for mutable values

type family Mutable a :: * -> *

class FreezeThaw a where

Methods

freeze :: PrimMonad m => Mutable a (PrimState m) -> m a

thaw :: PrimMonad m => a -> m (Mutable a (PrimState m))

unsafeFreeze :: PrimMonad m => Mutable a (PrimState m) -> m a

unsafeThaw :: PrimMonad m => a -> m (Mutable a (PrimState m))

Point

module OpenCV.Core.Types.Point

Size

module OpenCV.Core.Types.Size

Scalar

data Scalar

A 4-element vector with 64 bit floating point elements

The type Scalar is widely used in OpenCV to pass pixel values.

OpenCV Sphinx doc

class ToScalar a where

Methods

toScalar :: a -> Scalar

class FromScalar a where

Methods

fromScalar :: Scalar -> a

Rect

module OpenCV.Core.Types.Rect

RotatedRect

data RotatedRect

Rotated (i.e. not up-right) rectangles on a plane

Each rectangle is specified by the center point (mass center), length of each side (represented by Size2f) and the rotation angle in degrees.

OpenCV Sphinx doc

mkRotatedRect

Arguments

:: (IsPoint2 point2 CFloat, IsSize size CFloat)

=> point2 CFloat
Rectangle mass center

-> size CFloat
Width and height of the rectangle

-> Float
The rotation angle (in degrees). When the angle is 0, 90, 180, 270 etc., the rectangle becomes an up-right rectangle.

-> RotatedRect

rotatedRectCenter :: RotatedRect -> Point2f

Rectangle mass center

rotatedRectSize :: RotatedRect -> Size2f

Width and height of the rectangle

rotatedRectAngle :: RotatedRect -> Float

The rotation angle (in degrees)

When the angle is 0, 90, 180, 270 etc., the rectangle becomes an up-right rectangle.

rotatedRectBoundingRect :: RotatedRect -> Rect2i

The minimal up-right rectangle containing the rotated rectangle

rotatedRectPoints :: RotatedRect -> (Point2f, Point2f, Point2f, Point2f)

TermCriteria

data TermCriteria

Termination criteria for iterative algorithms

OpenCV Sphinx doc

mkTermCriteria

Arguments

:: Maybe Int
Optionally the maximum number of iterations/elements.

-> Maybe Double
Optionally the desired accuracy.

-> TermCriteria

Range

data Range

A continuous subsequence (slice) of a sequence

The type is used to specify a row or a column span in a matrix (Mat) and for many other purposes. mkRange a b is basically the same as a:b in Matlab or a..b in Python. As in Python, start is an inclusive left boundary of the range and end is an exclusive right boundary of the range. Such a half-opened interval is usually denoted as [start, end).

OpenCV Sphinx doc

mkRange :: Int32 -> Int32 -> Range

wholeRange :: Range

KeyPoint

data KeyPoint

Data structure for salient point detectors

OpenCV Sphinx doc

data KeyPointRec

Constructors

KeyPointRec

Fields

kptPoint :: !(V2 Float)

Coordinates of the keypoints.

kptSize :: !Float

Diameter of the meaningful keypoint neighborhood.

kptAngle :: !Float

Computed orientation of the keypoint (-1 if not applicable); it's in [0,360) degrees and measured relative to image coordinate system, ie in clockwise.

kptResponse :: !Float

The response by which the most strong keypoints have been selected. Can be used for the further sorting or subsampling.

kptOctave :: !Int32

Octave (pyramid layer) from which the keypoint has been extracted.

kptClassId :: !Int32

Object class (if the keypoints need to be clustered by an object they belong to).

mkKeyPoint :: KeyPointRec -> KeyPoint

keyPointAsRec :: KeyPoint -> KeyPointRec

DMatch

data DMatch

Class for matching keypoint descriptors: query descriptor index, train descriptor index, train image index, and distance between descriptors

OpenCV Sphinx Doc

data DMatchRec

Constructors

DMatchRec

Fields

dmatchQueryIdx :: !Int32

Query descriptor index.

dmatchTrainIdx :: !Int32

Train descriptor index.

dmatchImgIdx :: !Int32

Train image index.

dmatchDistance :: !Float

mkDMatch :: DMatchRec -> DMatch

dmatchAsRec :: DMatch -> DMatchRec

Matrix

module OpenCV.Core.Types.Mat

module OpenCV.Core.Types.Matx

Vec

module OpenCV.Core.Types.Vec

Exception

module OpenCV.Exception

Algorithm

class Algorithm a m where

Methods

algorithmClearState :: a -> m ()

algorithmIsEmpty :: a -> m Bool

Polymorphic stuff

class WithPtr a

Perform an IO action with a pointer to the C equivalent of a value

Minimal complete definition

withPtr

class FromPtr a

Types of which a value can be constructed from a pointer to the C equivalent of that value

Used to wrap values created in C.

Minimal complete definition

fromPtr

class CSizeOf a

Information about the storage requirements of values in C

This class assumes that the type a is merely a symbol that corresponds with a type in C.

Minimal complete definition

cSizeOf

class PlacementNew a

Copy source to destination using C++'s placement new feature

Minimal complete definition

placementNew, placementDelete