OpenCV.Core.Types

Point

2D types

data Point2i

2D point with integer coördinates

OpenCV Sphinx doc

data Point2f

2D point with 32 bit floating point coördinates

OpenCV Sphinx doc

data Point2d

2D point with 64 bit floating point coördinates

OpenCV Sphinx doc

class ToPoint2i a where

Methods

toPoint2i :: a -> Point2i

class FromPoint2i a where

Methods

fromPoint2i :: Point2i -> a

class ToPoint2f a where

Methods

toPoint2f :: a -> Point2f

class FromPoint2f a where

Methods

fromPoint2f :: Point2f -> a

class ToPoint2d a where

Methods

toPoint2d :: a -> Point2d

class FromPoint2d a where

Methods

fromPoint2d :: Point2d -> a

3D types

data Point3i

3D point with integer coördinates

OpenCV Sphinx doc

data Point3f

3D point with 32 bit floating point coördinates

OpenCV Sphinx doc

data Point3d

3D point with 64 bit floating point coördinates

OpenCV Sphinx doc

class ToPoint3i a where

Methods

toPoint3i :: a -> Point3i

class FromPoint3i a where

Methods

fromPoint3i :: Point3i -> a

class ToPoint3f a where

Methods

toPoint3f :: a -> Point3f

class FromPoint3f a where

Methods

fromPoint3f :: Point3f -> a

class ToPoint3d a where

Methods

toPoint3d :: a -> Point3d

class FromPoint3d a where

Methods

fromPoint3d :: Point3d -> a

newtype Vec3f

3D vector with 32 bit floating point coordinates

Constructors

Vec3f

Fields

unVec3f :: ForeignPtr (C Vec3f)

class FromVec3f a where

Methods

fromVec3f :: Vec3f -> a

4D types

newtype Vec4i

4D vector with integer coordinates

Constructors

Vec4i

Fields

unVec4i :: ForeignPtr (C Vec4i)

class ToVec4i a where

Methods

toVec4i :: a -> Vec4i

class FromVec4i a where

Methods

fromVec4i :: Vec4i -> a

Size

data Size2i

Size of an image or rectangle with integer values

OpenCV Sphinx doc

data Size2f

Size of an image or rectangle with 32 bit floating point values

OpenCV Sphinx doc

class ToSize2i a where

Methods

toSize2i :: a -> Size2i

class FromSize2i a where

Methods

fromSize2i :: Size2i -> a

class ToSize2f a where

Methods

toSize2f :: a -> Size2f

class FromSize2f a where

Methods

fromSize2f :: Size2f -> a

Scalar

data Scalar

A 4-element vector with 64 bit floating point elements

The type Scalar is widely used in OpenCV to pass pixel values.

OpenCV Sphinx doc

class ToScalar a where

Methods

toScalar :: a -> Scalar

class FromScalar a where

Methods

fromScalar :: Scalar -> a

Rect

data Rect

2D rectangles

OpenCV Sphinx doc

mkRect

Arguments

:: V2 Int32
top left

-> V2 Int32
size

-> Rect

rectTopLeft :: Rect -> Point2i

The top-left corner

rectBottomRight :: Rect -> Point2i

The bottom-right corner

rectSize :: Rect -> Size2i

Size (width, height) of the rectangle

rectArea :: Rect -> Int32

Area (width*height) of the rectangle

rectContains :: ToPoint2i point2i => point2i -> Rect -> Bool

Checks whether the rectangle contains the point

RotatedRect

data RotatedRect

Rotated (i.e. not up-right) rectangles on a plane

Each rectangle is specified by the center point (mass center), length of each side (represented by Size2f) and the rotation angle in degrees.

OpenCV Sphinx doc

mkRotatedRect

Arguments

:: (ToPoint2f point2f, ToSize2f size2f)

=> point2f
Rectangle mass center

-> size2f
Width and height of the rectangle

-> Float
The rotation angle (in degrees). When the angle is 0, 90, 180, 270 etc., the rectangle becomes an up-right rectangle.

-> RotatedRect

rotatedRectCenter :: RotatedRect -> Point2f

Rectangle mass center

rotatedRectSize :: RotatedRect -> Size2f

Width and height of the rectangle

rotatedRectAngle :: RotatedRect -> Float

The rotation angle (in degrees)

When the angle is 0, 90, 180, 270 etc., the rectangle becomes an up-right rectangle.

rotatedRectBoundingRect :: RotatedRect -> Rect

The minimal up-right rectangle containing the rotated rectangle

rotatedRectPoints :: RotatedRect -> (Point2f, Point2f, Point2f, Point2f)

TermCriteria

data TermCriteria

Termination criteria for iterative algorithms

OpenCV Sphinx doc

mkTermCriteria

Arguments

:: Maybe Int
Optionally the maximum number of iterations/elements.

-> Maybe Double
Optionally the desired accuracy.

-> TermCriteria

Range

data Range

A continuous subsequence (slice) of a sequence

The type is used to specify a row or a column span in a matrix (Mat) and for many other purposes. mkRange a b is basically the same as a:b in Matlab or a..b in Python. As in Python, start is an inclusive left boundary of the range and end is an exclusive right boundary of the range. Such a half-opened interval is usually denoted as [start, end).

OpenCV Sphinx doc

mkRange :: Int32 -> Int32 -> Range

wholeRange :: Range

KeyPoint

data KeyPoint

Data structure for salient point detectors

OpenCV Sphinx doc

data KeyPointRec

Constructors

KeyPointRec

Fields

kptPoint :: !(V2 Float)

Coordinates of the keypoints.

kptSize :: !Float

Diameter of the meaningful keypoint neighborhood.

kptAngle :: !Float

Computed orientation of the keypoint (-1 if not applicable); it's in [0,360) degrees and measured relative to image coordinate system, ie in clockwise.

kptResponse :: !Float

The response by which the most strong keypoints have been selected. Can be used for the further sorting or subsampling.

kptOctave :: !Int32

Octave (pyramid layer) from which the keypoint has been extracted.

kptClassId :: !Int32

Object class (if the keypoints need to be clustered by an object they belong to).

mkKeyPoint :: KeyPointRec -> KeyPoint

keyPointAsRec :: KeyPoint -> KeyPointRec

DMatch

data DMatch

Class for matching keypoint descriptors: query descriptor index, train descriptor index, train image index, and distance between descriptors

OpenCV Sphinx Doc

data DMatchRec

Constructors

DMatchRec

Fields

dmatchQueryIdx :: !Int32

Query descriptor index.

dmatchTrainIdx :: !Int32

Train descriptor index.

dmatchImgIdx :: !Int32

Train image index.

dmatchDistance :: !Float

mkDMatch :: DMatchRec -> DMatch

dmatchAsRec :: DMatch -> DMatchRec

Matrix

module OpenCV.Core.Types.Mat

Exception

module OpenCV.Exception

Algorithm

class Algorithm a m where

Methods

algorithmClearState :: a -> m ()

algorithmIsEmpty :: a -> m Bool

Polymorphic stuff

type family PointT dim depth :: *

Calculates a point type given the dimension and element type

Equations

PointT 2 Int32 = Point2i

PointT 2 Float = Point2f

PointT 2 Double = Point2d

PointT 2 CFloat = Point2f

PointT 2 CDouble = Point2d

PointT 3 Int32 = Point3i

PointT 3 Float = Point3f

PointT 3 Double = Point3d

PointT 3 CFloat = Point3f

PointT 3 CDouble = Point3d

class WithPtr a

Perform an IO action with a pointer to the C equivalent of a value

Minimal complete definition

withPtr

class FromPtr a

Types of which a value can be constructed from a pointer to the C equivalent of that value

Used to wrap values created in C.

Minimal complete definition

fromPtr

class CSizeOf a

Information about the storage requirements of values in C

This class assumes that the type a is merely a symbol that corresponds with a type in C.

Minimal complete definition

cSizeOf

class PlacementNew a

Copy source to destination using C++'s placement new feature

Minimal complete definition

placementNew, placementDelete