OpenCV.ImgCodecs

Documentation

data ImreadMode

Constructors

ImreadUnchanged

ImreadGrayscale

ImreadColor

ImreadAnyDepth

ImreadAnyColor

ImreadLoadGdal

imdecode :: ImreadMode -> ByteString -> Mat (S '[D, D]) D D

Reads an image from a buffer in memory.

The function reads an image from the specified buffer in the memory. If the buffer is too short or contains invalid data, the empty matrix/image is returned.

OpenCV Sphinx doc

imdecodeM :: PrimMonad m => ImreadMode -> ByteString -> m (Mut (Mat (S '[D, D]) D D) (PrimState m))

data OutputFormat

Constructors

OutputBmp

OutputExr

OutputHdr Bool
Compression (run length encoding)

OutputJpeg JpegParams

OutputJpeg2000

OutputPng PngParams

OutputPxm Bool
Binary

OutputSunras

OutputTiff

OutputWebP Int
Quality [1..100], > 100 == lossless

data JpegParams

Constructors

JpegParams

Fields

jpegParamQuality :: Int

\[0..100\]

jpegParamProgressive :: Bool

jpegParamOptimize :: Bool

jpegParamRestartInterval :: Word16

jpegParamLumaQuality :: Int

jpegParamChromaQuality :: Int

defaultJpegParams :: JpegParams

data PngStrategy

Constructors

PngStrategyDefault

PngStrategyFiltered

PngStrategyHuffmanOnly

PngStrategyRLE

PngStrategyFixed

data PngParams

Constructors

PngParams

Fields

pngParamCompression :: Int

pngParamStrategy :: PngStrategy

pngParamBinaryLevel :: Bool

defaultPngParams :: PngParams

imencode :: OutputFormat -> Mat shape channels depth -> CvExcept ByteString

Encodes an image into a memory buffer.

WARNING: This function is not thread safe!

OpenCV Sphinx doc

imencodeM :: PrimMonad m => OutputFormat -> Mut (Mat shape channels depth) (PrimState m) -> CvExceptT m ByteString

Encodes an image into a memory buffer.

See imencode