OpenCV.Internal.Core.Types

Scalar

newtype Scalar

A 4-element vector with 64 bit floating point elements

The type Scalar is widely used in OpenCV to pass pixel values.

OpenCV Sphinx doc

Constructors

Scalar

Fields

unScalar :: ForeignPtr (C Scalar)

newScalar :: V4 CDouble -> IO Scalar

class ToScalar a where

Minimal complete definition

toScalar

Methods

toScalar :: a -> Scalar

class FromScalar a where

Minimal complete definition

fromScalar

Methods

fromScalar :: Scalar -> a

RotatedRect

newtype RotatedRect

Rotated (i.e. not up-right) rectangles on a plane

Each rectangle is specified by the center point (mass center), length of each side (represented by Size2f) and the rotation angle in degrees.

OpenCV Sphinx doc

Constructors

RotatedRect

Fields

unRotatedRect :: ForeignPtr (C RotatedRect)

newRotatedRect

Arguments

:: (IsPoint2 point2 CFloat, IsSize size CFloat)

=> point2 CFloat
Rectangle mass center

-> size CFloat
Width and height of the rectangle

-> CFloat
The rotation angle (in degrees). When the angle is 0, 90, 180, 270 etc., the rectangle becomes an up-right rectangle.

-> IO RotatedRect

TermCriteria

newtype TermCriteria

Termination criteria for iterative algorithms

OpenCV Sphinx doc

Constructors

TermCriteria

Fields

unTermCriteria :: ForeignPtr (C TermCriteria)

newTermCriteria

Arguments

:: Maybe Int
Optionally the maximum number of iterations/elements.

-> Maybe Double
Optionally the desired accuracy.

-> IO TermCriteria

Range

newtype Range

A continuous subsequence (slice) of a sequence

The type is used to specify a row or a column span in a matrix (Mat) and for many other purposes. mkRange a b is basically the same as a:b in Matlab or a..b in Python. As in Python, start is an inclusive left boundary of the range and end is an exclusive right boundary of the range. Such a half-opened interval is usually denoted as [start, end).

OpenCV Sphinx doc

Constructors

Range

Fields

unRange :: ForeignPtr (C Range)

newRange

Arguments

:: Int32
Inclusive start

-> Int32
Exlusive end

-> IO Range

newWholeRange :: IO Range

Special Range value which means "the whole sequence" or "the whole range"

Polygons

withPolygons :: forall a point2. IsPoint2 point2 Int32 => Vector (Vector (point2 Int32)) -> (Ptr (Ptr (C Point2i)) -> IO a) -> IO a

withArrayPtr :: forall a b. (WithPtr a, CSizeOf (C a), PlacementNew (C a)) => Vector a -> (Ptr (C a) -> IO b) -> IO b

Perform an action with a temporary pointer to an array of values

The input values are placed consecutively in memory using the PlacementNew mechanism.

This function is intended for types which are not managed by the Haskell runtime, but by a foreign system (such as C).

The pointer is not guaranteed to be usuable outside the scope of this function. The same warnings apply as for withForeignPtr.

Orphan instances

PlacementNew C'Scalar

Methods

placementNew :: Ptr C'Scalar -> Ptr C'Scalar -> IO ()

placementDelete :: Ptr C'Scalar -> IO ()