OpenCV.Video.MotionAnalysis

BackgroundSubtractor

class BackgroundSubtractor a m where

Minimal complete definition

bgSubApply, getBackgroundImage

Methods

bgSubApply :: a -> Double -> Mat (S '[h, w]) channels depth -> m (Mat (S '[h, w]) (S 1) (S Word8))

getBackgroundImage :: a -> m (Mat (S '[h, w]) channels depth)

Background subtractors

data BackgroundSubtractorMOG2 s

data BackgroundSubtractorKNN s

newBackgroundSubtractorKNN

Arguments

:: PrimMonad m

=> Maybe Int32
Length of the history.

-> Maybe Double
Threshold on the squared distance between the pixel and the sample to decide whether a pixel is close to that sample. This parameter does not affect the background update.

-> Maybe Bool
If True, the algorithm will detect shadows and mark them. It decreases the speed a bit, so if you do not need this feature, set the parameter to False.

-> m (BackgroundSubtractorKNN (PrimState m))

newBackgroundSubtractorMOG2

Arguments

:: PrimMonad m

=> Maybe Int32
Length of the history.

-> Maybe Double
Threshold on the squared Mahalanobis distance between the pixel and the model to decide whether a pixel is well described by the background model. This parameter does not affect the background update.

-> Maybe Bool
If True, the algorithm will detect shadows and mark them. It decreases the speed a bit, so if you do not need this feature, set the parameter to False.

-> m (BackgroundSubtractorMOG2 (PrimState m))